+++ src/app/garage-emitter/garage-emitter.ts
@@ -1,6 +1,6 @@
 import type { Car } from '../../services/garage-service/types';
 
-import { Emitter } from '../../shared/event-emitter/event-emitter';
+import { Emitter } from '../../shared/emitter/emitter';
 
 export interface GaragePageEvents {
   'garage:create-hundred': never;
